@@ -7,27 +7,36 @@
     "test": "jest"
   },
   "dependencies": {
-    "@exponent/ex-navigation": "^2.2.1",
+    "@exponent/ex-navigation": "^2.9.1",
     "@exponent/vector-icons": "^2.0.3",
+    "babel-cli": "^6.26.0",
+    "babel-plugin-transform-class-properties": "^6.24.1",
+    "babel-preset-es2015": "^6.24.1",
     "babel-preset-react-native-stage-0": "^1.0.1",
     "immutable": "^3.8.1",
     "jwt-decode": "^2.2.0",
     "lodash": "^4.17.4",
-    "react": "^15.4.2",
-    "react-native": "^0.41.2",
-    "react-native-camera": "^0.6.0",
-    "react-native-code-push": "^1.17.0-beta",
+    "react": "16.0.0-alpha.6",
+    "react-cloud-vision-api": "^0.3.0",
+    "react-native": "0.43.3",
+    "react-native-auth0": "^1.2.1",
+    "react-native-camera": "^0.9.0",
+    "react-native-fetch-blob": "^0.10.8",
     "react-native-keyboard-aware-scrollview": "^1.1.4",
+    "react-native-maps": "^0.17.0",
+    "react-native-open-settings": "^1.0.1",
     "react-redux": "^5.0.1",
     "redux": "^3.6.0",
     "redux-immutable": "^3.0.9",
     "redux-logger": "^2.7.4",
+    "redux-persist-immutable": "^4.3.1",
     "redux-thunk": "^2.1.0",
     "tcomb-form-native": "^0.6.4"
   },
   "devDependencies": {
     "babel-jest": "18.0.0",
     "babel-preset-react-native": "1.9.1",
+    "babel-preset-stage-0": "^6.24.1",
     "jest": "18.0.0",
     "react-test-renderer": "15.4.1"
   },
