@@ -1,16 +1,17 @@
 package com.adsnap;
 
 import android.app.Application;
-import android.util.Log;
 
 import com.facebook.react.ReactApplication;
-import com.microsoft.codepush.react.CodePush;
+import com.airbnb.android.react.maps.MapsPackage;
 import com.lwansbrough.RCTCamera.RCTCameraPackage;
-import com.facebook.react.ReactInstanceManager;
+import com.auth0.react.A0Auth0Package;
+import com.RNFetchBlob.RNFetchBlobPackage;
 import com.facebook.react.ReactNativeHost;
 import com.facebook.react.ReactPackage;
 import com.facebook.react.shell.MainReactPackage;
 import com.facebook.soloader.SoLoader;
+import com.opensettings.OpenSettingsPackage;
 
 import java.util.Arrays;
 import java.util.List;
@@ -18,11 +19,6 @@ import java.util.List;
 public class MainApplication extends Application implements ReactApplication {
 
   private final ReactNativeHost mReactNativeHost = new ReactNativeHost(this) {
-
-    @Override
-    protected String getJSBundleFile() {
-      return CodePush.getJSBundleFile();
-    }
 
     @Override
     public boolean getUseDeveloperSupport() {
@@ -33,8 +29,11 @@ public class MainApplication extends Application implements ReactApplication {
     protected List<ReactPackage> getPackages() {
       return Arrays.<ReactPackage>asList(
           new MainReactPackage(),
-            new CodePush(getResources().getString(R.string.reactNativeCodePush_androidDeploymentKey), getApplicationContext(), BuildConfig.DEBUG),
-            new RCTCameraPackage()
+            new MapsPackage(),
+          new A0Auth0Package(),
+          new RNFetchBlobPackage(),
+          new RCTCameraPackage(),
+          new OpenSettingsPackage()
       );
     }
   };
