@@ -15,9 +15,9 @@ import {
 import { Provider } from 'react-redux';
 import store from './src/redux/store';
 import AppViewContainer from './src/modules/AppViewContainer';
-import codePush from "react-native-code-push";
+// import codePush from "react-native-code-push";
 
-@codePush
+// @codePush
 export default class adsnap extends Component {
   render() {
     return (
