@@ -1,16 +1,23 @@
-export const BACK_BUTTON = require('../../images/icon_back.png');
-export const QR = require('../../images/icon_qrcode.png');
-export const CAMERA = require('../../images/icon_camera.png');
-export const VOUCHER = require('../../images/icon_voucher.png');
-export const LOGO_WHITE = require('../../images/logo_white.png');
-export const LOGO_COLORED = require('../../images/logo_colored.png');
+export const BACK_BUTTON = require("../../images/icon_back.png");
+export const QR = require("../../images/icon_qrcode.png");
+export const CAMERA = require("../../images/icon_camera.png");
+export const VOUCHER = require("../../images/icon_voucher.png");
+export const LOGO_WHITE = require("../../images/logo_white.png");
+export const LOGO_COLORED = require("../../images/logo_colored.png");
+export const CHALLENGES = require("../../images/icon_challenges.png");
+export const SAD = require("../../images/icon_sad.png");
+export const GEOLOCATION = require('../../images/icon_geolocation.png');
+export const MALE_ACTIVE = require("../../images/icon_male_selected.png");
+export const MALE_INACTIVE = require("../../images/icon_male_unselected.png");
+export const FEMALE_ACTIVE = require("../../images/icon_female_selected.png");
+export const FEMALE_INACTIVE = require("../../images/icon_female_unselected.png");
+export const OTHER_ACTIVE = require("../../images/icon_other_selected.png");
+export const OTHER_INACTIVE = require("../../images/icon_other_unselected.png");
+export const PROFILE = require("../../images/icon_profile.png");
+export const ICON_AT_BLACK = require("../../images/icon_at_black.png");
 
-export const MALE_ACTIVE = require('../../images/icon_male_selected.png');
-export const MALE_INACTIVE = require('../../images/icon_male_unselected.png');
-export const FEMALE_ACTIVE = require('../../images/icon_female_selected.png');
-export const FEMALE_INACTIVE = require('../../images/icon_female_unselected.png');
-export const OTHER_ACTIVE = require('../../images/icon_other_selected.png');
-export const OTHER_INACTIVE = require('../../images/icon_other_unselected.png');
+export const LIST = require("../../images/icon_map.png");
+export const MAP= require("../../images/icon_list.png");
 
 /* STOCK PHOTOS */
-export const WELCOME_BG = require('../../images/bg_welcome.png');
+export const WELCOME_BG = require("../../images/bg_welcome.png");
