@@ -1,8 +1,7 @@
 import AppView from './AppView';
-import {connect} from 'react-redux';
+import { connect } from 'react-redux';
 
 export default connect(
-  state => ({
-      
+  state => ({      
   })
 )(AppView);
