@@ -1,18 +1,17 @@
-import {StyleSheet, Dimensions} from 'react-native'
+import { StyleSheet, Dimensions } from 'react-native'
 import * as COLORS from '../../constants/colors';
 import * as COMMON_STYLES from '../../constants/commonStyles';
 
 export default styles = StyleSheet.create({
-    table:{
-        flexDirection: 'column'
-    },
-    tileWrapper: {
-        flexDirection: 'row',
-        flex:1,
-    },
-    tile: {
-        // backgroundColor: '#CCC',
-        margin: 3,
-
-    }
+	table: {
+		flexDirection: 'row',
+		flexWrap: 'wrap'
+	},
+	tileWrapper: {
+		flexDirection: 'row',
+		flex: 1,
+	},
+	tile: {
+		margin: 3,
+	}
 })
