@@ -5,6 +5,7 @@ import { toJS } from 'immutable';
 export default connect(
   state => ({
     deals: state.getIn(['deals', 'deals']).toJS(),
-    colors: state.getIn(['deals', 'colors']).toJS()
+    activeDeal: state.getIn(['deals', 'activeDeal']).toJS(),
+    colors: state.getIn(['deals', 'colors']).toJS(),
   })
 )(DealsView);
