@@ -1,38 +1,38 @@
-import * as RedeemState from './RedeemState';
 import * as COLORS from '../../constants/colors';
 import React from 'react';
 import {
-    Image,
-    Text,
-    View,
-    Dimensions,
-    ScrollView
+  Image,
+  Text,
+  View,
+  Dimensions,
+  ScrollView,
+  Clipboard
 } from 'react-native';
-import * as ICONS from '../../constants/icons';
 import Container from '../../components/Container';
 import KeyValue from '../../components/KeyValue';
 import RectButton from '../../components/RectButton';
-import Tile from '../../components/Tile';
 import * as COMMON_STYLES from '../../constants/commonStyles';
 
 import styles from './RedeemStyles';
-
 
 export default class RedeemView extends React.Component {
     static route = {
         navigationBar: {
             title: 'Redeem',
         }
+    };
+
+    copyToClipboard() {
+      console.log('copied to clipboard');
+      Clipboard.setString(this.props.deal.code);
     }
-    //TODO: fetch code
-    //TODO: popup on remove
+
     render() {
-        console.log(this.props)
         return (
             <Container loading={!this.props.deal}>
                 <ScrollView>
                     <View style={styles.codeWrapper}>
-                        <Image style={styles.backgroundImage} source={{ uri: this.props.deal.campaignImgUrl }}>
+                        <Image style={styles.backgroundImage} source={this.props.deal.campaignImgUrl}>
                             {/*<View style={styles.coloredOverlay} />
                             <View style={styles.codeBox}><Text style={styles.codeText}>{this.props.route.params.deal.code}</Text></View>*/}
                         </Image>
@@ -62,24 +62,24 @@ export default class RedeemView extends React.Component {
                     </Text>*/}
                         <View style={styles.buttonWrapper}>
                             <RectButton
-                                onPress={() => { }}
-                                text={'Trade'}
+                                onPress={() => this.copyToClipboard()}
+                                text={'Copy to clipboard'}
                                 width={COMMON_STYLES.BUTTON_WIDTH(Dimensions)}
                                 height={COMMON_STYLES.BUTTON_HEIGHT}
                                 textColor={COLORS.WHITE}
                                 backgroundColor={COLORS.DARK_GREY}
                                 borderColor={COLORS.BLACK} />
                         </View>
-                        <View style={styles.buttonWrapper}>
-                            <RectButton
-                                onPress={() => { }}
-                                text={'Remove Discount'}
-                                width={COMMON_STYLES.BUTTON_WIDTH(Dimensions)}
-                                height={COMMON_STYLES.BUTTON_HEIGHT}
-                                textColor={COLORS.WHITE}
-                                backgroundColor={COLORS.DARK_GREY}
-                                borderColor={COLORS.BLACK} />
-                        </View>
+                        {/*<View style={styles.buttonWrapper}>*/}
+                            {/*<RectButton*/}
+                                {/*onPress={() => { }}*/}
+                                {/*text={'Remove Discount'}*/}
+                                {/*width={COMMON_STYLES.BUTTON_WIDTH(Dimensions)}*/}
+                                {/*height={COMMON_STYLES.BUTTON_HEIGHT}*/}
+                                {/*textColor={COLORS.WHITE}*/}
+                                {/*backgroundColor={COLORS.DARK_GREY}*/}
+                                {/*borderColor={COLORS.BLACK} />*/}
+                        {/*</View>*/}
                     </View>
                 </ScrollView>
             </Container >
